split `checkAPIformat.lua` into multiple files
